fixes, scoring, linting (#1)

+++ commands/ffmpeg_sampler.go
@@ -6,13 +6,14 @@ import (
 	"fmt"
 	"os"
 	"os/exec"
+	"strconv"
 
 	"github.com/rs/zerolog"
 	"github.com/terranvigil/dynamic-crf/model"
 )
 
 type FfmpegSampler struct {
-	log              zerolog.Logger
+	logger           zerolog.Logger
 	sourcePath       string
 	sampleOutputPath string
 	fps              float64
@@ -21,7 +22,7 @@ type FfmpegSampler struct {
 
 func NewFfmpegSampler(log zerolog.Logger, sourcePath string, samplePath string, fps float64, scenes []model.Scene) *FfmpegSampler {
 	return &FfmpegSampler{
-		log:              log,
+		logger:           log,
 		sourcePath:       sourcePath,
 		sampleOutputPath: samplePath,
 		fps:              fps,
@@ -40,36 +41,42 @@ func (s *FfmpegSampler) Run(ctx context.Context) error {
 		}
 		return total
 	}()
-	s.log.Info().Msgf("running ffmpeg sampler, creating sample of %.2fs duration", totalDur)
+	s.logger.Info().Msgf("running ffmpeg sampler, creating sample of %.2fs duration", totalDur)
 	tempPaths := []string{}
 
 	for i, scene := range s.scenes {
-		s.log.Info().Msgf("scene #%d:, start: %.2fs, dur: %.2fs", i+1, scene.StartPTSSec, scene.Duration)
+		if err := func() error {
+			s.logger.Info().Msgf("scene #%d:, start: %.2fs, dur: %.2fs", i+1, scene.StartPTSSec, scene.Duration)
 
-		f, err := os.CreateTemp("", "smpl_*.ts")
-		if err != nil {
-			return fmt.Errorf("failed to create temp file for encoding sample, err: %w, message: %s", err, stderr.String())
-		}
-		defer os.Remove(f.Name())
-		tempPaths = append(tempPaths, f.Name())
+			f, err := os.CreateTemp("", "smpl_*.ts")
+			if err != nil {
+				return fmt.Errorf("failed to create temp file for encoding sample, err: %w, message: %s", err, stderr.String())
+			}
+			defer os.Remove(f.Name())
+			tempPaths = append(tempPaths, f.Name())
 
-		args := []string{
-			"-hide_banner",
-			// TODO add this back in to target keyframe before start time
-			//"-ss", strconv.Itoa(r[0]),
-			"-i", s.sourcePath,
-			"-ss", fmt.Sprintf("%.2f", scene.StartPTSSec),
-			"-frames:v", fmt.Sprintf("%d", int(s.fps*scene.Duration)),
-			"-c:v", "copy",
-			"-an",
-			"-y", f.Name(),
-		}
-		s.log.Debug().Msgf("ffmpeg args: %v", args)
+			args := []string{
+				"-hide_banner",
+				// TODO add this back in to target keyframe before start time
+				// "-ss", strconv.Itoa(r[0]),
+				"-i", s.sourcePath,
+				"-ss", fmt.Sprintf("%.2f", scene.StartPTSSec),
+				"-frames:v", strconv.Itoa(int(s.fps * scene.Duration)),
+				"-c:v", "copy",
+				"-an",
+				"-y", f.Name(),
+			}
+			s.logger.Debug().Msgf("ffmpeg args: %v", args)
 
-		cmd := exec.CommandContext(ctx, "ffmpeg", args...)
-		cmd.Stderr = &stderr
-		if err = cmd.Run(); err != nil {
-			return fmt.Errorf("ffmpeg sample creation %s failed, err: %w, message: %s", f.Name(), err, stderr.String())
+			cmd := exec.CommandContext(ctx, "ffmpeg", args...)
+			cmd.Stderr = &stderr
+			if err = cmd.Run(); err != nil {
+				return fmt.Errorf("ffmpeg sample creation %s failed, err: %w, message: %s", f.Name(), err, stderr.String())
+			}
+
+			return nil
+		}(); err != nil {
+			return err
 		}
 	}
